// pop-web/src/app/shared/openapi/model/genomic-variant-clinical-relevance-choices.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type GenomicVariantClinicalRelevanceChoices = 'pathogenic' | 'likely_pathogenic' | 'uncertain_significance' | 'ambiguous' | 'likely_benign' | 'benign';

export const GenomicVariantClinicalRelevanceChoices = {
    Pathogenic: 'pathogenic' as GenomicVariantClinicalRelevanceChoices,
    LikelyPathogenic: 'likely_pathogenic' as GenomicVariantClinicalRelevanceChoices,
    UncertainSignificance: 'uncertain_significance' as GenomicVariantClinicalRelevanceChoices,
    Ambiguous: 'ambiguous' as GenomicVariantClinicalRelevanceChoices,
    LikelyBenign: 'likely_benign' as GenomicVariantClinicalRelevanceChoices,
    Benign: 'benign' as GenomicVariantClinicalRelevanceChoices
};