// pop-web/src/app/shared/openapi/model/genomic-variant-assessment-choices.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type GenomicVariantAssessmentChoices = 'present' | 'absent' | 'no-call' | 'indeterminate';

export const GenomicVariantAssessmentChoices = {
    Present: 'present' as GenomicVariantAssessmentChoices,
    Absent: 'absent' as GenomicVariantAssessmentChoices,
    NoCall: 'no-call' as GenomicVariantAssessmentChoices,
    Indeterminate: 'indeterminate' as GenomicVariantAssessmentChoices
};